+++ packages/terra-list/src/terra-dev-site/doc/guides/MultiSelectList.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import List, { Item, Utils } from 'terra-list/lib/index';
-import Placeholder from 'terra-doc-template/lib/Placeholder';
+import {Placeholder} from '@cerner/terra-docs';
 import classNames from 'classnames/bind';
 import mockData from './mock-data/mock-select';
 import styles from '../example/ListDocCommon.module.scss';