+++ packages/terra-list/src/terra-dev-site/doc/guides/SectionWithSubsection2.jsx
@@ -5,7 +5,7 @@ import List, {
   Subsection,
   Utils,
 } from 'terra-list';
-import Placeholder from 'terra-doc-template/lib/Placeholder';
+import {Placeholder} from '@cerner/terra-docs';
 import classNames from 'classnames/bind';
 import mockData from './mock-data/mock-section-sub';
 import styles from '../example/ListDocCommon.module.scss';