+++ packages/terra-markdown/src/terra-dev-site/test/markdown/scss.test.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import Markdown from '../../../../lib/Markdown';
-import markdown from './scss.md';
-
-const ScssTest = () => (
-  <Markdown src={markdown} />
-);
-
-export default ScssTest;