+++ packages/terra-table/src/terra-dev-site/doc/common/FixedNode.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Placeholder from 'terra-doc-template/lib/Placeholder';
+import {Placeholder} from '@cerner/terra-docs';
 import styles from './FixedNode.module.scss';
 
 const cx = classNames.bind(styles);