+++ packages/terra-action-header/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Replaced references to terra-doc-template with terra-docs
+
 ## 2.56.0 - (August 18, 2020)
 
 * Changed